Renamed the tests for acollin for better info
